Add review system with flagging in the frontend

+++ backend/src/main/java/edu/ap/gosmartlib/dto/reviews/ReviewDetailDTO.java
@@ -14,6 +14,6 @@ public record ReviewDetailDTO(
         String text,
         LocalDate reviewDate,
         ReviewStatus reviewStatus,
-        int rating,
+        float rating,
         int flagCount
 ) {}
